fixed bug with model attributes not loading. Moved model attribute from BookController to the HomeController
